rft: generic invocation samples
