+++ generic/default/go-server/conf/dubbogo.yml
@@ -8,9 +8,15 @@ dubbo:
     dubbo:
       name: dubbo
       port: 20000
+    triple:
+      name: tri
+      port: 20001
   provider:
     registry: zk
     services:
       UserProvider:
         protocol: dubbo
-        interface: org.apache.dubbo.UserProvider
+        interface: org.apache.dubbo.samples.UserProvider
+      UserProviderTriple:
+        protocol: triple
+        interface: org.apache.dubbo.samples.UserProviderTriple
