+++ generic/default/go-server/pkg/user_response.go
@@ -1,0 +1,9 @@
+package pkg
+
+type UserResponse struct {
+	Users []*User
+}
+
+func (u *UserResponse) JavaClassName() string {
+	return "org.apache.dubbo.UserResponse"
+}
